remove default value, move DB model to module dbmodel
